update channel status page
This commit updates how the channel status is displayed on the channel status
page. Instead of each channel status word having its own column and then
displaying check marks if that word is set, it now shows the status as a list
of strings in a single column. This makes the table fit better on the page and
gives space for the info column which was otherwise being smooshed in the last
column.