add a nocache decorator to json views
When viewing the stream page on Firefox, it tends to cache every single request
which can result in it saving *millions* of small cache files over a long
period of time and using up all the inodes on the monitoring machines.

Hopefully, by setting the correct headers on the responses we can make it so
Firefox doesn't cache these results, although I haven't tested this yet.